requirements.txt
msc-pyparser>=1.2.1
google-re2>=1.0
